Breaking out list.getUniqueValue into its own file
